+++ src/main/java/net/minecraft/client/gui/GuiVideoSettings.java
@@ -207,7 +207,7 @@ public class GuiVideoSettings extends GuiScreenOF
 
         this.drawString(this.fontRendererObj, s, 2, this.height - 10, 8421504);
         String s2 = "Minecraft 1.8.9";
-        int i = this.fontRendererObj.getStringWidth(s2);
+        int i = (int) this.fontRendererObj.getStringWidth(s2);
         this.drawString(this.fontRendererObj, s2, this.width - i - 2, this.height - 10, 8421504);
         super.drawScreen(mouseX, mouseY, partialTicks);
         this.tooltipManager.drawTooltips(mouseX, mouseY, this.buttonList);
