+++ src/main/java/net/optifine/shaders/gui/GuiShaderOptions.java
@@ -120,7 +120,7 @@ public class GuiShaderOptions extends GuiScreenOF
         {
             FontRenderer fontrenderer = Config.getMinecraft().fontRendererObj;
 
-            for (int i = fontrenderer.getStringWidth(": " + Lang.getOff()) + 5; fontrenderer.getStringWidth(s) + i >= btnWidth && s.length() > 0; s = s.substring(0, s.length() - 1))
+            for (int i = (int) (fontrenderer.getStringWidth(": " + Lang.getOff()) + 5); fontrenderer.getStringWidth(s) + i >= btnWidth && s.length() > 0; s = s.substring(0, s.length() - 1))
             {
                 ;
             }
