+++ src/main/java/net/optifine/shaders/gui/GuiShaders.java
@@ -345,7 +345,7 @@ public class GuiShaders extends GuiScreenOF
 
         this.drawCenteredString(this.fontRendererObj, this.screenTitle + " ", this.width / 2, 15, 16777215);
         String s = "OpenGL: " + Shaders.glVersionString + ", " + Shaders.glVendorString + ", " + Shaders.glRendererString;
-        int i = this.fontRendererObj.getStringWidth(s);
+        int i = (int) this.fontRendererObj.getStringWidth(s);
 
         if (i < this.width - 5)
         {
